refactor : userId push하는 hook 구현 및 router로 위치 변경
