+++ src/pages/FineBook/index.tsx
@@ -1,18 +1,8 @@
-import { userState } from '@/store/userState';
-import { useEffect } from 'react';
-import { useRecoilValue } from 'recoil';
 import Calendar from '@/components/@common/Calendar';
 import DetailFine from './DetailFine';
 import * as Style from './styles';
 
 const FineBook = () => {
-  const user = useRecoilValue(userState);
-  useEffect(() => {
-    if (user.userId) {
-      window.dataLayer.push({ user_id: user.userId });
-    }
-  }, [user.userId]);
-
   return (
     <Style.Layout>
       <Calendar cellType="Mark" />
